[Refactor] EmbeddedRedis 오류 수정 및 redis config 파일 통합 - 통합 테스트 중 Embedded Redis가 캐싱되어 있는데 계속 redis를 다시 만들어 start를 하면 오류가 발생함. - 원인은 어플리케이션 컨텍스트가 캐싱되어 있기 떄문이다. 그렇기에 Redis start 에러는 skip 처리
