+++ src/test/java/kr/or/connect/reservation/ProfileControllerTest.java
@@ -1,5 +1,6 @@
 package kr.or.connect.reservation;
 
+import kr.or.connect.reservation.domain.config.RedisConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,7 +11,8 @@ import org.springframework.http.ResponseEntity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+        classes = {RedisConfig.class})
 class ProfileControllerTest {
 
     @LocalServerPort
