+++ src/test/resources/application.properties
@@ -1,35 +1,35 @@
-# H2 인메모리 데이터베이스 설정 (MySQL 모드로 동작시키려면 MODE=MySQL 옵션 추가)
+# H2 \uC778\uBA54\uBAA8\uB9AC \uB370\uC774\uD130\uBCA0\uC774\uC2A4 \uC124\uC815 (MySQL \uBAA8\uB4DC\uB85C \uB3D9\uC791\uC2DC\uD0A4\uB824\uBA74 MODE=MySQL \uC635\uC158 \uCD94\uAC00)
 spring.datasource.driver-class-name=org.h2.Driver
 spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;MODE=MySQL;DATABASE_TO_LOWER=TRUE
 
 spring.datasource.username=sa
 spring.datasource.password=
 
-# colorful 로그찍기
+# colorful \uB85C\uADF8\uCC0D\uAE30
 spring.output.ansi.enabled=always
 
-# H2 콘솔 활성화 (필요시)
+# H2 \uCF58\uC194 \uD65C\uC131\uD654 (\uD544\uC694\uC2DC)
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
-# JPA 설정
+# JPA \uC124\uC815
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.properties.hibernate.hbm2ddl.auto=create
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.type.descriptor.sql.trace=true
+#spring.jpa.show-sql=true
+#spring.jpa.properties.hibernate.format_sql=true
+#spring.jpa.properties.hibernate.type.descriptor.sql.trace=true
 
-# resources/data 의 sql 파일을 실행하기 위한 설정
+# resources/data \uC758 sql \uD30C\uC77C\uC744 \uC2E4\uD589\uD558\uAE30 \uC704\uD55C \uC124\uC815
 # spring.sql.init.mode=always
 
-# Redis 설정
+# Redis \uC124\uC815
 spring.data.redis.host=localhost
 spring.data.redis.port=6379
 
-# insert sql 문 로그 출력
+# insert sql \uBB38 \uB85C\uADF8 \uCD9C\uB825
 #logging.level.org.springframework.jdbc.datasource.init.ScriptUtils=DEBUG
 #logging.level.org.springframework.jdbc.datasource.init.DataSourceInitializer=DEBUG
 #logging.level.org.hibernate.type.descriptor.sql.BasicBinder=trace
 
-# p6spy로 sql 로그 출력
+# p6spy\uB85C sql \uB85C\uADF8 \uCD9C\uB825
 decorator.datasource.p6spy.enable-logging=true
